requirements.txt
locust
fastapi
requests
grpcio 
grpcio-tools
fastapi
uvicorn[standard]
protoletariat